update types and use components
